style: rename models folder to types
the `models` folder name was introduced to differentiate it from the parent `types` folder. With the parent `types` folder no longer existing, this folder has been renamed for clarity
