+++ app/components/Socials/Socials.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styles from './Socials.module.css';
 import SectionHeader from '../SectionHeader/SectionHeader';
-import { Social } from '../../models/Social';
+import { Social } from '../../types/Social';
 import { icons } from './Icons';
 
 interface Props {
