+++ app/types/Artist.ts
