+++ app/types/EventBrief.ts
